Refactoring: separate package for trace generation
